+++ reckon_crm/reckon_crm/doctype/reckon_crm_settings/reckon_crm_settings.py
@@ -1,0 +1,7 @@
+from frappe.model.document import Document
+from reckon_crm.services.settings import validate_settings
+
+
+class ReckonCRMSettings(Document):
+    def validate(self):
+        validate_settings(self)